PROD-244 Follow 응답 핸들러를 분리한다
