+++ packages/fedify/src/federation.ts
@@ -1,9 +1,10 @@
 import { createFederation, MemoryKvStore } from '@fedify/fedify';
 import { Accept, Create, Follow, Reject, Undo } from '@fedify/vocab';
 import { resolveConfiguredLocalInstance } from '@kosmo/core/local-instance';
+import { handleInboundAcceptFollow } from './inbound-accept-follow';
 import { handleInboundCreate } from './inbound-create';
 import { handleInboundFollow, handleInboundUndo } from './inbound-follow';
-import { handleInboundAccept, handleInboundReject } from './inbound-follow-response';
+import { handleInboundRejectFollow } from './inbound-reject-follow';
 import { ensureDrizzleLocalProfileActor } from './local-actor-store';
 import { createLocalProfilePerson } from './local-profile-person';
 import { resolveLocalActorIdentifierByHandle } from './webfinger';
@@ -61,8 +62,8 @@ federation
 
 federation
   .setInboxListeners('/ap/actor/{identifier}/inbox', '/inbox')
-  .on(Accept, handleInboundAccept)
+  .on(Accept, handleInboundAcceptFollow)
   .on(Create, handleInboundCreate)
   .on(Follow, handleInboundFollow)
-  .on(Reject, handleInboundReject)
+  .on(Reject, handleInboundRejectFollow)
   .on(Undo, handleInboundUndo);
